Generate alevin summary report

Generate a report summarizing the main aspects of an alevin quantification run. The report generation assumes that alevin has been run with the --dumpFeatures flag to generate the necessary output files.

alevinQCReport(
  baseDir,
  sampleId,
  outputFile,
  outputDir = "./",
  outputFormat = NULL,
  showCode = FALSE,
  forceOverwrite = FALSE,
  knitrProgress = FALSE,
  quiet = FALSE,
  ignorePandoc = FALSE,
  customCBList = list(),
  ...
)

Arguments

baseDir
Path to the output directory from the alevin run (should be the directory containing the alevin directory).

sampleId
Sample ID, will be used to set the title for the report.

outputFile
File name of the output report. The file name extension must be either .html or .pdf, and consistent with the value of outputFormat.

outputDir
Path to the output directory where the report will be generated.

outputFormat
The format of the output report. Either "html_document" or "pdf_document". The file name extension of outputFile must be consistent with this choice.

showCode
Logical, whether to display the R code in the report.

forceOverwrite
Logical, whether to force overwrite an existing report with the same name in the output directory.

knitrProgress
Logical, whether to display the progress of knitr when generating the report.

quiet
Logical, whether to show progress messages.

ignorePandoc
Logical, determines what to do if pandoc or pandoc-citeproc is missing (if Sys.which("pandoc") or Sys.which("pandoc-citeproc") returns ""). If ignorePandoc is TRUE, only a warning is given. The figures will be generated, but not the final report. If ignorePandoc is FALSE (default), the execution stops immediately.

customCBList
Named list with custom set(s) of barcodes to provide summary statistics/plots for, in addition to the whitelists generated by alevin.

...
Other arguments that will be passed to rmarkdown::render.

Value

Generates a summary report in the outputDir directory, and returns (invisibly) the name of the generated report.

Details

When the function is called, a .Rmd template file will be copied into the output directory, and rmarkdown::render will be called to generate the final report. If there is already a .Rmd file with the same name in the output directory, the function will raise an error and stop, to avoid overwriting the existing file. The reason for this behaviour is that the copied template in the output directory will be deleted once the report is generated.

Examples

alevinQCReport(baseDir = system.file("extdata/alevin_example_v0.14",
                                     package = "alevinQC"),
               sampleId = "example", outputFile = "alevinReport.html",
               outputDir = tempdir(), forceOverwrite = TRUE)

#> Reading Alevin output files...

#> Generating summary tables...

#> Generating knee plot...

#> Generating barcode collapsing plot...

#> Generating quantification summary plot...

#> Generating knee plot for nbr genes...